update wsl
rename the classes
