+++ pymic/net_run_wsl/wsl_mumford_shah.py
@@ -2,24 +2,20 @@
 from __future__ import print_function, division
 import logging
 import numpy as np
-import random
 import torch
-import torchvision.transforms as transforms
-from pymic.io.nifty_dataset import NiftyDataset
 from pymic.loss.seg.util import get_soft_label
 from pymic.loss.seg.util import reshape_prediction_and_ground_truth
 from pymic.loss.seg.util import get_classwise_dice
 from pymic.loss.seg.mumford_shah import MumfordShahLoss
-from pymic.net_run.agent_seg import SegmentationAgent
-from pymic.net_run_wsl.wsl_em import WSL_EntropyMinimization
+from pymic.net_run_wsl.wsl_abstract import WSLSegAgent
 from pymic.util.ramps import sigmoid_rampup
 
-class WSL_MumfordShah(WSL_EntropyMinimization):
+class WSLMumfordShah(WSLSegAgent):
     """
     Training and testing agent for semi-supervised segmentation
     """
     def __init__(self, config, stage = 'train'):
-        super(WSL_MumfordShah, self).__init__(config, stage)
+        super(WSLMumfordShah, self).__init__(config, stage)
 
     def training(self):
         class_num   = self.config['network']['class_num']
